requirements.txt
pytest>=7.0.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
google-api-python-client>=2.80.0